Most of login page completed
But index simply displays the username
You need to have a MySQL database called users, with a table called
users, with at least one <username, password HASH> pair stored, in your
XAMPP for this to work. The hash I'm using now is just PHP's
password_hash(password, PASSWORD_BCRYPT), but I think this is based on
DES, so I'll have to go back and make it stronger.