Users folder (very rudimentary database for users)
The users folder must be copied into the "\mysql\data\" folder of your
XAMPP. Also, you have to make sure on your XAMPP that MySQL and Tomcat
(Catalina) are activated. (There are mysql start and catalina start .bat
files in the root directory of your XAMPP if you find that these are
deactivated.)
The only (username, password) that works right now is
(username=stephenattech, password=GodIsGood414). This is not my actual
password for any real system, of course! Again, the hash I'm using (for
now) is password_hash(password, PASSWORD_BCRYPT), but I'll have to
change this some time closer to when we put the final project together.